backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9
pydantic==2.8.2
sentence-transformers==3.0.1
faiss-cpu==1.8.0
numpy==1.26.4
google-generativeai==0.7.2
pymupdf==1.24.9
python-docx==1.1.2
python-dotenv==1.0.1
